+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 2.1.0 (Roadmap)
-- Minimum supported version of Puppet bumped to 5.0.0.
+- Minimum supported version of Puppet bumped to 5.0.
+- Minimum Ruby version bumped to 2.3.
+- Official support for Puppet version 7.
 
 ### 2.0.1
 - Check for existence of executables for dependency module retrieval.